pyproject.toml
[tool.poetry]
name = "appium-setup"
version = "0.1.0"
description = ""
authors = ["Andrey Stoletov <stoletov314@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "4.9.0"
pytest = "^7.4.3"
chromedriver-autoinstaller = "^0.6.2"
chromedriver = "^2.24.1"
appium-python-client = "2.9.0"
requests = "^2.31.0"
pytest-ordering = "^0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

